+++ files/default/tests/minitest/ruby_test.rb
@@ -19,10 +19,10 @@
 
 require File.expand_path('../support/helpers', __FILE__)
 
-describe 'postgresql::ruby' do
+describe 'rackspace_postgresql::ruby' do
   include Helpers::Postgresql
 
   it 'installs the pg gem in Chefs ruby environment' do
-    assert Gem::Specification.all_names.grep("pg-.*")
+    assert Gem::Specification.all_names.grep('pg-.*')
   end
 end