+++ files/default/tests/minitest/support/helpers.rb
@@ -18,12 +18,17 @@
 #
 
 module Helpers
+  #######
+  # Some test helpers
   module Postgresql
+    #######
+    # Postgresql helpers
+    #
+    # Utilizes minitests and chef's shellouts
     require 'chef/mixin/shell_out'
     include Chef::Mixin::ShellOut
     include MiniTest::Chef::Assertions
     include MiniTest::Chef::Context
     include MiniTest::Chef::Resources
-
   end
 end