+++ recipes/contrib.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: postgresql
+# Cookbook Name:: rackspace_postgresql
 # Recipe:: contrib
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,11 +15,11 @@
 # limitations under the License.
 #
 
-include_recipe "postgresql::server"
+include_recipe 'rackspace_postgresql::server'
 
 # Install the PostgreSQL contrib package(s) from the distribution,
 # as specified by the node attributes.
-node['postgresql']['contrib']['packages'].each do |pg_pack|
+node['rackspace_postgresql']['contrib']['packages'].each do |pg_pack|
 
   package pg_pack
 
@@ -27,8 +27,8 @@ end
 
 # Install PostgreSQL contrib extentions into the template1 database,
 # as specified by the node attributes.
-if (node['postgresql']['contrib'].attribute?('extensions'))
-  node['postgresql']['contrib']['extensions'].each do |pg_ext|
+if node['rackspace_postgresql']['contrib'].attribute?('extensions')
+  node['rackspace_postgresql']['contrib']['extensions'].each do |pg_ext|
     bash "install-#{pg_ext}-extension" do
       user 'postgres'
       code <<-EOH
@@ -36,7 +36,7 @@ if (node['postgresql']['contrib'].attribute?('extensions'))
       EOH
       action :run
       ::Chef::Resource.send(:include, Opscode::PostgresqlHelpers)
-      not_if {extension_installed?(pg_ext)}
+      not_if { extension_installed?(pg_ext) }
     end
   end
 end