+++ recipes/server_redhat.rb
@@ -4,7 +4,9 @@
 #
 # Author:: Joshua Timberman (<joshua@opscode.com>)
 # Author:: Lamont Granquist (<lamont@opscode.com>)
+# Author:: Matthew Thode (<matt.thode@rackspace.com>)
 # Copyright 2009-2011, Opscode, Inc.
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,54 +21,54 @@
 # limitations under the License.
 #
 
-include_recipe "postgresql::client"
+include_recipe 'rackspace_postgresql::client'
 
 # Create a group and user like the package will.
 # Otherwise the templates fail.
 
-group "postgres" do
+group 'postgres' do
   gid 26
 end
 
-user "postgres" do
-  shell "/bin/bash"
-  comment "PostgreSQL Server"
-  home "/var/lib/pgsql"
-  gid "postgres"
+user 'postgres' do
+  shell '/bin/bash'
+  comment 'PostgreSQL Server'
+  home '/var/lib/pgsql'
+  gid 'postgres'
   system true
   uid 26
-  supports :manage_home => false
+  supports manage_home: false
 end
 
-directory node['postgresql']['dir'] do
-  owner "postgres"
-  group "postgres"
+directory node['rackspace_postgresql']['dir'] do
+  owner 'postgres'
+  group 'postgres'
   recursive true
   action :create
 end
 
-node['postgresql']['server']['packages'].each do |pg_pack|
+node['rackspace_postgresql']['server']['packages'].each do |pg_pack|
 
   package pg_pack
 
 end
 
-template "/etc/sysconfig/pgsql/#{node['postgresql']['server']['service_name']}" do
-  source "pgsql.sysconfig.erb"
-  mode "0644"
-  notifies :restart, "service[postgresql]", :delayed
+template "/etc/sysconfig/pgsql/#{node['rackspace_postgresql']['server']['service_name']}" do
+  source 'pgsql.sysconfig.erb'
+  mode '0644'
+  notifies :restart, 'service[postgresql]', :delayed
 end
 
-unless platform_family?("suse")
+unless platform_family?('suse')
 
-  execute "/sbin/service #{node['postgresql']['server']['service_name']} initdb #{node['postgresql']['initdb_locale']}" do
-    not_if { ::FileTest.exist?(File.join(node['postgresql']['dir'], "PG_VERSION")) }
+  execute "/sbin/service #{node['rackspace_postgresql']['server']['service_name']} initdb #{node['rackspace_postgresql']['initdb_locale']}" do
+    not_if { ::FileTest.exist?(File.join(node['rackspace_postgresql']['dir'], 'PG_VERSION')) }
   end
 
 end
 
-service "postgresql" do
-  service_name node['postgresql']['server']['service_name']
-  supports :restart => true, :status => true, :reload => true
+service 'postgresql' do
+  service_name node['rackspace_postgresql']['server']['service_name']
+  supports restart: true, status: true, reload: true
   action [:enable, :start]
 end